+++ CHANGELOG.md
@@ -93,6 +93,7 @@ and this project adheres to
   from Rust into Go.
 - cosmwasm-vm: Upgrade Wasmer to 4.2.5; Bump `MODULE_SERIALIZATION_VERSION` to
   "v9". ([#1992])
+- cosmwasm-std: Rename `GovMsg::vote` to `GovMsg::option` ([#1999])
 
 [#1874]: https://github.com/CosmWasm/cosmwasm/pull/1874
 [#1876]: https://github.com/CosmWasm/cosmwasm/pull/1876
@@ -115,6 +116,7 @@ and this project adheres to
 [#1973]: https://github.com/CosmWasm/cosmwasm/pull/1973
 [#1977]: https://github.com/CosmWasm/cosmwasm/pull/1977
 [#1992]: https://github.com/CosmWasm/cosmwasm/pull/1992
+[#1999]: https://github.com/CosmWasm/cosmwasm/pull/1999
 
 ### Removed